Consolidate code for getting platform details into a single place
